sync table schema

+++ CarRental.DataAccess/DB/CarDB/CarModel.cs
@@ -7,7 +7,7 @@ namespace CarRental.DataAccess.DB.CarDB
 {
     public partial class CarModel
     {
-        public ulong Id { get; set; }
+        public uint Id { get; set; }
         public uint CarBrandId { get; set; }
         public string Name { get; set; }
     }
